* Renamed sdcard_archive_path to use camellias * Moved generation of a toast for sdcard failure and the sdcardArchivePath methods into the DeveloperActivity
